requirements.txt
Django>=4.2.0,<5.0.0
WeasyPrint>=59.0
Pillow>=10.0.0
psycopg2-binary>=2.9.5
django-crispy-forms>=2.0
